Added network comp and auth state management
Using Axios for making network calls with interceptors to handle header injection and error processing.
Using Redux for state management. Auth state is a managed store slice and can be accessed as typed state.

The api facility is designed to work with a strapi instance.

The following files are added/changed:

- Auth state store slice (AuthStoreSlice.tsx)
- Primnary redux store (Store.tsx)
- Network client as static utils (NetworkClient.tsx)
- Updated landing page to include mock/test buttons